make loginWith3rdParty global and show toast on error only
